refactor: dedup storybook ClientAPI stub and drop dead code
- Add shared withApiProvider() decorator (mocks/api-decorator.tsx),
  replacing the ClientAPI stub copy-pasted across 3 component stories
- Delete unused tooling/storybook decorators.tsx (divergent 4th stub)
- Remove dead createLogContents factory
- Revert redundant inspect-stories glob in packages/react storybook
  (central @tsmono/storybook host already covers them, with MSW wiring)
- Use createEvalHeader for ErrorState header instead of a flat literal
