+++ apps/inspect/src/components/DownloadButton.stories.tsx
@@ -1,34 +1,12 @@
-import type { Decorator, Meta, StoryObj } from "@storybook/react";
+import type { Meta, StoryObj } from "@storybook/react";
 
-import type { ClientAPI } from "../client/api/types";
-import { ApiProvider } from "../state/store";
+import { withApiProvider } from "../mocks/api-decorator";
 
 import { DownloadButton } from "./DownloadButton";
 
-const stubApi: ClientAPI = {
-  get_log_dir: async () => undefined,
-  get_log_dir_handle: () => "",
-  get_logs: async () => ({ log_dir: "", files: [] }) as never,
-  get_log_root: async () => ({ logs: [] }) as never,
-  get_eval_set: async () => undefined,
-  get_flow: async () => undefined,
-  get_log_summaries: async () => [],
-  get_log_details: async () => ({ eval: {}, sampleSummaries: [] }) as never,
-  get_log_sample: async () => undefined,
-  client_events: async () => [],
-  download_file: async () => {},
-  open_log_file: async () => {},
-};
-
-const withApiProvider: Decorator = (Story) => (
-  <ApiProvider value={stubApi}>
-    <Story />
-  </ApiProvider>
-);
-
 const meta = {
   component: DownloadButton,
-  decorators: [withApiProvider],
+  decorators: [withApiProvider()],
   parameters: {
     layout: "padded",
   },
