+++ apps/inspect/src/mocks/api-decorator.tsx
@@ -1,0 +1,28 @@
+import type { Decorator } from "@storybook/react";
+
+import type { ClientAPI } from "../client/api/types";
+import { ApiProvider } from "../state/store";
+
+const stubApi: ClientAPI = {
+  get_log_dir: async () => undefined,
+  get_log_dir_handle: () => "",
+  get_logs: async () => ({ log_dir: "", files: [] }) as never,
+  get_log_root: async () => ({ logs: [] }) as never,
+  get_eval_set: async () => undefined,
+  get_flow: async () => undefined,
+  get_log_summaries: async () => [],
+  get_log_details: async () => ({ eval: {}, sampleSummaries: [] }) as never,
+  get_log_sample: async () => undefined,
+  client_events: async () => [],
+  download_file: async () => {},
+  open_log_file: async () => {},
+};
+
+/** Wraps a story in an <ApiProvider/> backed by a no-op stub API. */
+export const withApiProvider =
+  (overrides: Partial<ClientAPI> = {}): Decorator =>
+  (Story) => (
+    <ApiProvider value={{ ...stubApi, ...overrides }}>
+      <Story />
+    </ApiProvider>
+  );
