+++ packages/react/.storybook/main.ts
@@ -1,10 +1,7 @@
 import type { StorybookConfig } from "@storybook/react-vite";
 
 const config: StorybookConfig = {
-  stories: [
-    "../src/**/*.stories.@(ts|tsx)",
-    "../../../apps/inspect/src/**/*.stories.@(ts|tsx)",
-  ],
+  stories: ["../src/**/*.stories.@(ts|tsx)"],
   framework: "@storybook/react-vite",
 };
